+++ HISTORY.rst
@@ -3,6 +3,26 @@
 Release History
 ---------------
 
+1.0.8 (2025-07-04)
+++++++++++++++++++
+
+**Updates**
+
+- Add tests for image without paragraph. | [Dfop02](https://github.com/dfop02)
+- Add tests for bookmark without paragraph. | [Dfop02](https://github.com/dfop02)
+- Add tests for local image. | [Dfop02](https://github.com/dfop02)
+- Add tests for unbalanced table. | [Dfop02](https://github.com/dfop02)
+
+**Fixes**
+
+- Fix crash when there is bookmark without paragraph. | [Dfop02](https://github.com/dfop02) from [Issue](https://github.com/dfop02/html4docx/issues/21)
+- Fix crash when there is image without paragraph. | [Dfop02](https://github.com/dfop02) from [Issue](https://github.com/dfop02/html4docx/issues/19)
+
+**New Features**
+
+- None
+
+
 1.0.7 (2025-06-17)
 ++++++++++++++++++